### 2025-05-12 #8191 @editorialbot
[PRE REVIEW]: SC2Tools: StarCraft II Toolset and Dataset API
**Submitting author:** @Kaszanas (<a href="http://orcid.org/0000-0003-3668-4638">Andrzej Białecki</a>)
**Repository:** https://github.com/Kaszanas/SC2Tools
**Branch with paper.md** (empty if default branch): draft_paper
**Version:** 1.0.0
