### 2025-05-12 7f8b351
Attempting to fix PDF generation, JOSS
