+++ .github/workflows/draft-pdf.yaml
@@ -4,6 +4,7 @@ on:
     paths:
       - article/**
       - .github/workflows/**
+      - paper.md
   workflow_dispatch: # Allows manual triggering of the workflow
 
 jobs:
@@ -18,7 +19,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: article/paper.md
+          paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02
         with:
@@ -26,4 +27,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: article/paper.pdf
+          path: paper.pdf
