### 2025-06-14 #8191 @Kaszanas
@crvernon I'll try to do that. I'm worried that reducing word count so heavily might botch the idea of pooling multiple repositories together.

You can imagine that describing each of the submodules in its own right can lead to additional word count.

If you have any solutions I'll be happy to implement them. One article per repository with a meta-article could be one of them, but that'd require major work.

I'll refer to the word count specifications and see what I can do.

> @Kaszanas while I am working on this, please reduce the word count of your paper to < 1000 words and add a statement of need section to your paper.  You can refer the the submission instructions to see an example.  Thanks.
