### 2025-06-21 #8191 @crvernon
@Kaszanas - I see your justification in the paper for the research software applications here with RL; however, I want to make sure our editorial board agrees that there is a clear research software link when there are ties to a specific game and the tools are less generalized.  I am also worried that finding reviewers for a specific use case may be challenging.  I will get back to your ASAP with a decision.  Thank you.
