### 2025-06-21 #8191 @Kaszanas
That's mostly aligned with the use case for the software. Essentially these are tools for working with StarCraft 2 that lead to multiple published works as a result of multiple of years of work.

But use of DatasetPreparator is a little more generic.

The software itself could be reviewed as any other. I think it's documented well. If the docs are not satisfying, I am willing to make changes.

Streamlining the data pipelines for research makes it worth it in my opinion.

> @editorialbot query scope
>
> @Kaszanas - I see your justification in the paper for the research software applications here with RL; however, I want to make sure our editorial board agrees that there is a clear research software link when there are ties to a specific game and the tools are less generalized.  I am also worried that finding reviewers for a specific use case may be challenging.  I will get back to your ASAP with a decision.  Thank you.
