### 2025-07-16 #8191 @mbobra
👋 @crvernon This is a little bit outside of my scope (both in terms of subject matter and programming language -- I don't know Go), so I'll pass. Thanks for the invite and please feel free to send another one my way.
