### 2025-07-18 #8191 @nkrusch
👋 @crvernon

I am not familiar with this area of research, though that is not an immediate issue. After reviewing the paper, I noticed the following:

1. **The paper is too long.** About 3000 words. There is much discussion dedicated to the describing software, which is seems out of place for a JOSS paper.

2. **The scientific purpose remains unclear to me.** E.g., What is the motivation for combining (specifically) reinforcement learning  and esports? What is special about this specific instance of a game, StarCraft 2?. I was hoping to find the statement of need (I suspect _"Potential impact"_ aims to be that section) to find the research motivation, but I am struggling.

3. **The top-level repository.** Seems it is just submodules. As a user, I would need some guidance at this top level. I see the comment _"Please refer to the respective submodule for its usage instructions."_ but find that unsatisfactory, because I presume this combination of submodules adds some value.

I would minimally request a major edit of the JOSS paper. I can revisit the editorial request after that.
