+++ paper.md
@@ -54,40 +54,16 @@ When such software is available, it is often hard to use for less technically pr
 
 Esports can be treated as a subset of gaming with additional requirements for players, such as tournament presence, organized play, training, and professionalization [@Formosa2022]. The study of esports is multidisciplinary in nature [@Brock2023Interdisciplinary;@Pizzo2022Interdisciplinary]. Due to the growing academic interest in the area of gaming and esports [@Yamanaka2021Review;@Reitman2020Review;@Tang2023Review;@Bialecki2024Review], it is key to provide tools for researchers capable of simplifying the process of acquiring large datasets efficiently, not only for authors interested in the area of computer science [@Ferenczi2024sc2_serizlizer;@SmerdovLoLDataset2021].
 
-In case of our implementation, we focus on solving problems within the StarCraft&nbsp;2 (SC2) infrastructure ecosystem. StarCraft&nbsp;2 is a real-time strategy game developed by Blizzard Entertainment. The game is known as one of the most prominent real-time strategy (RTS) esports titles [@Tyreal2020;Dal2020]. It is also characterized by its fast-paced gameplay and a high skill ceiling [@Migliore2021]. These attributes make for a great environment for testing various AI agents [@Ma2024LLMStarCraft2,@Vinyals2019;@Samvelyan2019SMAC;@Pearce2022CSGO]. Moreover, research in StarCraft&nbsp;2 is not limited to AI agents -- there are efforts to analyze the game from various perspectives and provide insights that can assist players in their gameplay [@URLSC2AICoach2022LLM;@URLSc2replaystats].
+We focus on solving problems within the StarCraft&nbsp;2 (SC2) infrastructure ecosystem. We solve the problem of ease of access to the data encoded in files with ".SC2Replay". StarCraft&nbsp;2 is a real-time strategy game developed by Blizzard Entertainment. The game is known as one of the most prominent real-time strategy (RTS) esports titles [@Tyreal2020;Dal2020]. It is also characterized by its fast-paced gameplay and a high skill ceiling [@Migliore2021]. These attributes make for a great environment for testing various AI agents [@Ma2024LLMStarCraft2,@Vinyals2019;@Samvelyan2019SMAC;@Pearce2022CSGO]. Moreover, research in StarCraft&nbsp;2 is not limited to AI agents -- there are efforts to analyze the game from various perspectives and provide insights that can assist players in their gameplay [@URLSC2AICoach2022LLM;@URLSc2replaystats].
 
-Our software collection is an open-source implementation of data extraction, and data interfacing tools for StarCraft&nbsp;2. We solve the problem of ease of access to the data encoded in files with ".SC2Replay" extension by using an open-source file extractor for proprietary MoPAQ (MPQ) file format. From this point on, we will refer to the MPQ files with the ".SC2Replay" extension as SC2Replay files.
-
-So far, our software was leveraged in preparation of major datasets: "SC2ReSet" [@Bialecki2022ReSetZenodo] and "SC2EGSet" [@Bialecki2023EGSetZenodo] with an accompanying peer-reviewed and published Data Descriptor article [@Bialecki2023SC2EGSet]. The output of our software was used in varying contexts indirectly. authors cited our work, some of them following the general flow of our exploration [@Kim2024]. Others put emphasis on statistical calculation within esports landscape [@Dupuy2024]. Finally authors describe our work in surveys of related work when working in another games [@Johar2024].
-
[…]
