+++ paper.md
@@ -50,7 +50,7 @@ Computer games as fully controlled simulated environments were used in major sci
 Despite heightened interest in research on gaming and esports, there are limited high-level libraries and tools made for rapid experimentation in some game titles. Researchers from various research disciplines have shown their interest in exploring gaming and esports, including: (1) psychology [@Campbell2018], (2) computer science [@Rashid2020;@Yuan2021ActorCritic], (3) education [@Jensen2024;@Jenny2021], (4) medical sciences [@Krarup2020109344], and others [@Holden2017Law;Nagorsky2020].
 The ability to tie these topics with the in-game data cannot be overstated.
 
-When such software is available, it is often hard to use for less technically proficient researchers. Data parsing libraries are prevalent in computer games, such as Counter-Strike [@AWPYXeno2020;ClarityGitHub], Rocket&nbsp;League [@URLBoxcars2016], Dota&nbsp;2 [@OpenDotaGitHub;@ClarityGitHub], and finally in StarCraft&nbsp;2 [@URLBlizzardS2ClientProto;@URLS2Prot2016;@GitHubSC2Reader].
+When such software is available, it is often hard to use for less technically proficient researchers. Data parsing libraries are prevalent in computer games, such as Counter-Strike [@AWPYXeno2020;@ClarityGitHub], Rocket&nbsp;League [@URLBoxcars2016], Dota&nbsp;2 [@OpenDotaGitHub;@ClarityGitHub], and finally in StarCraft&nbsp;2 [@URLBlizzardS2ClientProto;@URLS2Prot2016;@GitHubSC2Reader].
 
 Esports can be treated as a subset of gaming with additional requirements for players, such as tournament presence, organized play, training, and professionalization [@Formosa2022]. The study of esports is multidisciplinary in nature [@Brock2023Interdisciplinary;@Pizzo2022Interdisciplinary]. Due to the growing academic interest in the area of gaming and esports [@Yamanaka2021Review;@Reitman2020Review;@Tang2023Review;@Bialecki2024Review], it is key to provide tools for researchers capable of simplifying the process of acquiring large datasets efficiently, not only for authors interested in the area of computer science [@Ferenczi2024sc2_serizlizer;@SmerdovLoLDataset2021].
 
