### 2025-08-07 30ebbea
Further removals to hit the word count target
