### 2025-08-08 f487ef0
Adjusted the doc link, fixing figure rendering
