+++ paper.md
@@ -77,7 +77,7 @@ The SC2InfoExtractorGo as a submodule is a tool responsible for extracting the d
 
 One of our solutions, SC2_Datasets [@bialecki_2022_sc2datasets] interfaces with the JSON files produced by the SC2InfoExtractorGo [@Bialecki_2021_SC2InfoExtractorGo]. This includes all of the classes and methods required to load a single JSON, a collection of JSON files (representing a replaypack), and finally a way of loading an entire dataset (a collection of replaypacks). The pictorial representation of the "SC2_Datasets" functionality is presented on \autoref{fig:LoadingDataToPyTorch}.
 
-![Loading the output of the SC2InfoExtractorGo for machine learning and artificial intelligence use with "SC2_Datasets". \label{fig:LoadingDataToPyTorch}](SC2Tools_ExperimentWorkflow.pdf)
+![Loading the output of the SC2InfoExtractorGo for machine learning and artificial intelligence use with "SC2_Datasets". \label{fig:LoadingDataToPyTorch}](./article/SC2Tools_ExperimentWorkflow.pdf)
 
 Users have the ability to extend our solution and apply it to their data via the PyTorch [@PyTorch2019] and PyTorch Lightning [@PyTorch_Lightning_2019] interfaces.
 
@@ -93,7 +93,7 @@ The need for similar solutions is clear, output of our software was used in vary
 
 There exist many implementations built for the purpose of parsing replay files [@ZenodoSC2Reader]. These tools and libraries require expert programming skills to extract and interact with the resulting data. Many research approaches involve scientists that may not posses such expert knowledge in programming, but nonetheless interested in investigating esports (e.g., in psychology, biomechanics, social sciences and humanities -- SSH, and others) [@Kegelaers2025;@Dupuy2025;@Donghee2020]. Lowering the technical overhead needed to interact with in-game data can open gaming and esports to researchers with various non-technical backgrounds. Furthermore, integrating SSH scientists in the research process is not only a requirement in some funding programs, but also a practical necessity, if one aims to conduct socially responsible studies [@graf2019bringing;@sonetti2020only].
 
-Before introducing our software, users were bound to write their own tools extracting the data from StarCraft 2 replay files. Our solution outputs easy-to-use JSON files adhering to a specific, well-documented schema definition [https://sc2-datasets.readthedocs.io/en/latest/autoapi/index.html](https://sc2-datasets.readthedocs.io/en/latest/autoapi/index.html). Additionally, the data extraction toolset efficiently leverages modern multi-core processors (using Golang goroutines), making the process of data extraction faster. This has real implications on day-to-day research, as it allows for faster experimentation and iteration on one's methods. Finally, In the past, research conducted on StarCraft&nbsp;2 data has yielded fruitful ventures in online tooling [@URLSc2replaystats;@URLSpawningTool;@URLSC2Revealed;@URLAligulac]; and research [@Vinyals2019;@Ma2024LLMStarCraft2;@Samvelyan2019SMAC;@Ferenczi2024sc2_serizlizer].
[…]
