+++ paper.md
@@ -89,7 +89,7 @@ Main functionality of this software collection introduce a repeatable way of wor
 
 # Statement of Need
 
-The need for similar solutions is clear, output of our software was used in varying contexts, authors cited our work, some of them following the general flow of our exploration [@Kim2024;@Ferenczi2024sc2_serizlizer]. Within the intended user group, the software was created to assist with the process of StarCraft 2 data processing. Mainly, the software fulfilled the research needs of our team and other collaborating research teams, which led to processing and creating a dataset [@Bialecki2023SC2EGSet]. Additionally, an API interface was created to load and work with the data in PyTorch [@PyTorch2019] and PyTorch Lightning [@PyTorch_Lightning_2019].
+The need for similar solutions is clear, output of this software or related artifacts were used directly in varying contexts, the community cited our work, some authors following the general flow of our exploration [@Kim2024;@Ferenczi2024sc2_serizlizer]. Presented software was created to assist with the process of StarCraft 2 data processing. Mainly, the software fulfilled the research needs of our team, other collaborating research teams, and the scientific community as a whole. Additionally, as a result, the software was used to process and create a dataset shared openly [@Bialecki2023SC2EGSet]. Finally, an API interface was created to load and work with the data in PyTorch [@PyTorch2019] and PyTorch Lightning [@PyTorch_Lightning_2019].
 
 There exist many implementations built for the purpose of parsing replay files [@ZenodoSC2Reader]. These tools and libraries require expert programming skills to extract and interact with the resulting data. Many research approaches involve scientists that may not posses such expert knowledge in programming, but nonetheless interested in investigating esports (e.g., in psychology, biomechanics, social sciences and humanities -- SSH, and others) [@Kegelaers2025;@Dupuy2025;@Donghee2020]. Lowering the technical overhead needed to interact with in-game data can open gaming and esports to researchers with various non-technical backgrounds. Furthermore, integrating SSH scientists in the research process is not only a requirement in some funding programs, but also a practical necessity, if one aims to conduct socially responsible studies [@graf2019bringing;@sonetti2020only].
 
