### 2025-08-24 #8191 @nkrusch
Greetings 👋 @martinmodrak, @DiegoAscanio, @rusu24edward,

would you be interested in reviewing this submission for JOSS?
