### 2025-08-29 #8889 @nkrusch
Hello @Kaszanas, 👋

this will be the review thread so we will have all conversations under this issue from now on.

From the editorial bot checks, I see two issues that need immediate attention.

1. The [bibliography check](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3235321075) has a couple missing and invalid entries. Please resolve the bibliography issues.

2. I have noted [the license issue](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3235321158). It comes from the top-level repository not having a license. I also see the submodules use two different licenses. I have reviewed the conversation with the @crvernon during the pre-review, but did not find how/if this was ever resolved. I will check with the editorial board (no action is needed from your part at this time).
