### 2025-08-30 #8889 @nkrusch
Hi @DiegoAscanio,

I appreciate the quick turn-around and identifying the major issues.

**Regarding issues 1 and 2**: As these are about deficiencies of the software, I recommend you open these as issues against the target repository. This way we can ensure proper tracking and completion of fixes.

The paper-related issues are sufficient to mention in this comment thread.
