### 2025-08-30 #8889 @nkrusch
Side comment:
> Publish a versioned GitHub release and archive it with a DOI (e.g., Zenodo); cite that DOI in the paper.

Once the review tasks are completed, the final step will be to publish and archive.
@Kaszanas, please hold off from doing this until then. I will inform you when the time comes.
