### 2025-09-09 Kaszanas/SC2Tools#1 @nkrusch
Add installation and quickstart instructions
_(This issue is directly raised from the [review comments of @DiegoAscanio](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3238594209) for [JOSS review #8889](https://github.com/openjournals/joss-reviews/issues/8889))_

Add a top-level README.md with a clear Quickstart section.
Include exact commands such as:

```
git clone --recurse-submodules https://github.com/OWNER/SC2Tools.git
cd SC2Tools
git submodule update --init --recursive
```

Document any required build / install steps (e.g., Go build, pip requirements, system packages).
Provide at least one runnable example (Docker preferred) showing the full pipeline:

```
# hypothetical Docker example — adapt to the published image/name
docker run --rm -v $(pwd)/replays:/replays kaszanas/sc2infoextractor:latest   -i /replays/sample.SC2Replay -o /replays/sample.json
```
Include a one-line description of what each command does and how to verify success (expected files/outputs).
