### 2025-09-09 Kaszanas/SC2Tools#2 @nkrusch
End-to-end examples and reproducibility
_(This issue is directly raised from the [review comments of @DiegoAscanio](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3238594209) for [JOSS review #8889](https://github.com/openjournals/joss-reviews/issues/8889))_

- Include one or two working examples (scripts or Jupyter notebooks) that demonstrate an end-to-end workflow using either included sample replays or links to public replays.
- Examples must show:
  - the input (replay file or URL),
  - the exact commands run (CLI or Docker),
  - the produced outputs (JSON snippets or small printed examples),
  - and a short snippet showing how to load the produced JSON into common ML frameworks (e.g., PyTorch Dataset).
- Preferably provide a runnable notebook (e.g., examples/quickstart.ipynb) that a reviewer can open and run in Binder / Colab / locally.
