### 2025-09-26 #8889 @nkrusch
Hi @martinmodrak,

just checking up on status. I request that you post the following text as a comment, to generate your reviewer checklist.

There are already a few ongoing open issues -- summarized in [this comment](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3238594209).

At this time I would encourage you to start your checks, and see if that raises additional issues that were not covered by the previous comments.
