### 2025-10-05 #8889 @Kaszanas
> Hi @Kaszanas, just checking on status.
>
> I believe we are waiting for repository enhancements and updates to the paper. Do you have an estimate of when these changes would be ready?

I have delayed the changes as per the earlier suggestion. The second review seems to be in the works as I can't see it posted above.

Should I start making changes, or would you like for me to wait for all the reviews to be available @nkrusch

I'd like to limit the potential overlap between what will be done and potential comments which are not submitted as of now.
