### 2025-11-22 57b67cb
Added missing DOIs
