+++ article/paper.bib
@@ -115,7 +115,7 @@
   year      = {2021},
   publisher = {Zenodo},
   version   = {1.1.1},
-  doi       = {10.5281/zenodo.zenodo.4733264},
+  doi       = {10.5281/zenodo.4733264},
   url       = {https://doi.org/10.5281/zenodo.4733264}
 }
 @misc{Bialecki_2021_SC2InfoExtractorGo,
@@ -266,6 +266,7 @@
   eprint        = {2312.11865},
   archiveprefix = {arXiv},
   primaryclass  = {cs.AI},
+  doi           = {10.48550/arXiv.2312.11865},
   url           = {https://arxiv.org/abs/2312.11865}
 }
 @misc{Samvelyan2019SMAC,
@@ -275,6 +276,7 @@
   eprint        = {1902.04043},
   archiveprefix = {arXiv},
   primaryclass  = {cs.LG},
+  doi           = {10.48550/arXiv.1902.04043},
   url           = {https://arxiv.org/abs/1902.04043}
 }
 @article{Rashid2020,
@@ -290,6 +292,8 @@
   month      = {01},
   articleno  = {178},
   numpages   = {51},
+  doi        = {10.5555/3455716.3455894},
+  url        = {https://dl.acm.org/doi/abs/10.5555/3455716.3455894},
   keywords   = {multi-agent coordination, multi-agent learning, reinforcement learning}
 }
 @software{bialecki_2022_sc2datasets,
@@ -311,6 +315,7 @@
   eprint        = {2104.06655},
   archiveprefix = {arXiv},
   primaryclass  = {cs.AI},
+  doi           = {10.48550/arXiv.2104.06655},
   url           = {https://arxiv.org/abs/2104.06655}
 }
 @misc{Ferenczi2024sc2_serizlizer,
@@ -320,6 +325,7 @@
   eprint        = {2410.08659},
   archiveprefix = {arXiv},
   primaryclass  = {cs.LG},
+  doi           = {10.48550/arXiv.2410.08659},
   url           = {https://arxiv.org/abs/2410.08659}
 }
 @article{Kim2024,
@@ -351,13 +357,14 @@
 @article{Johar2024,
   title     = {{Modelling Player Skills in Rocket League through a Behavioural Pattern Mining Approach}},
   issn      = {3006-8894},
-  url       = {http://dx.doi.org/10.70023/piqm243},
   doi       = {10.70023/piqm243},
   journal   = {PatternIQ Mining},
   publisher = {Sahara Digital Publications},
   author    = {Johar,  Sayyed},
   year      = {2024},
-  month     = {02}
+  month     = {02},
+  doi       = {10.70023/piqm243},
+  url       = {http://dx.doi.org/10.70023/piqm243}
 }
 @inproceedings{Smerdov2019,
   author    = {Smerdov, Anton and Burnaev, Evgeny and Somov, Andrey},
@@ -388,8 +395,7 @@
   keywords  = {Human-Computer Interaction (cs.HC), Artificial Intelligence (cs.AI), Machine Learning (cs.LG), FOS: Computer and information sciences, FOS: Computer and information sciences},
   title     = {{Collection and Validation of Psychophysiological Data from Professional and Amateur Players: a Multimodal eSports Dataset}},
   publisher = {arXiv},
-  year      = {2020},
-  copyright = {arXiv.org perpetual, non-exclusive license}
+  year      = {2020}
 }
 @article{Tyreal2020,
   author  = {Tyreal Yizhou Qian and James Jianhui Zhang and Jerred Junqi Wang and John Hulland},
@@ -435,7 +441,8 @@
   pages     = {},
   publisher = {Curran Associates, Inc.},
   title     = {{PyTorch: An Imperative Style, High-Performance Deep Learning Library}},
[…]
