### 2025-11-22 e77b172
Fixed Dal2020 citation in text
