### 2026-01-05 04eb30c
Initial changes for #1
