### 2026-01-05 79f9f70
Refined README with the docker example
