### 2026-01-07 Kaszanas/SC2Tools#2 @nkrusch
By inspection, this still seems open.
