### 2026-01-07 #8889 @nkrusch
@Kaszanas

Hope you feel better soon.

I have checked the two technical issues, and they still need some attention from you. The other reviewer comments were about the paper. The technical issues and paper updates should be finished before re-engaging the reviewers.

Let me know when there are updates. It is my priority to move this submission toward completion in the next weeks.
