### 2026-01-09 Kaszanas/SC2Tools#2 @Kaszanas
> By inspection, this still seems open.

Correct, I will be pushing these changes in the next commits.
