### 2026-01-09 Kaszanas/SC2Tools#1 @Kaszanas
> [@Kaszanas](https://github.com/Kaszanas) Thanks! I checked the updated quick start on my end. There are still a few issues.
>
> 1. Docker command
>          [SC2Tools/README.md](https://github.com/Kaszanas/SC2Tools/blob/79f9f70ddd9669300b64ef50e5010e42ee1b8dd1/README.md?plain=1#L43-L45)
>            Lines 43 to 45
>          in
>          [79f9f70](/Kaszanas/SC2Tools/commit/79f9f70ddd9669300b64ef50e5010e42ee1b8dd1)
>               docker run -it --rm \
>               -v "${PWD}\processing\data":/app/processing/data \
>               kaszanas/datasetpreparator:latest \
>
>
>
>    I am running macOS, and had to change L44 to:
>    ```
>    -v "${PWD}/processing/data":/app/processing/data --platform=linux/amd64 \
>    ```
>    1. The Windows-style path `\` becomes `/` on unix hosts
>    2. The pre-built image does not launch on arm machines, but I can force amd by specifying `platform`
> 2. Python command
>
>
>
>          [SC2Tools/README.md](https://github.com/Kaszanas/SC2Tools/blob/79f9f70ddd9669300b64ef50e5010e42ee1b8dd1/README.md?plain=1#L46-L51)
>            Lines 46 to 51
>          in
>          [79f9f70](/Kaszanas/SC2Tools/commit/79f9f70ddd9669300b64ef50e5010e42ee1b8dd1)
>               python sc2egset_pipeline.py \
>               --input_dir processing/data/replays \
>               --output_dir processing/data/output \
>               --maps_path processing/maps \
>               --n_processes 4 \
>               --force_overwrite True
>
>
>    The argument names should be `input_path` and `output_path` (not `dir`).
>    However, since those directories do not exist, the command terminates.
>    ("Error: Invalid value for '--input_path': Directory 'processing/data/replays' does not exist.")
>    I stopped here.
>
> There are different ways to solve these (up to you)
>
> * The readme could be explicit that certain OS/architecture is assumed; or the docker command could be made to account for the differences (give different command/OS).
> * For the second, Python has [`makedirs`](https://docs.python.org/3/library/os.html#os.makedirs), which can be used to create the missing directories.

Thank you for spotting these, I will make adjustments.

>    However, since those directories do not exist, the command terminates.
>    ("Error: Invalid value for '--input_path': Directory 'processing/data/replays' does not exist.")
>    I stopped here.

The `input_path` by default should be pre-populated with replays so it needs to exist, otherwise the processing will run but nothing will happen. I can make the software create the directories and issue a warning message.
