### 2026-01-09 be60e1d
Improvements as suggested in #1
