### 2026-01-09 34b2f4a
Added makefile target, added examples/
