### 2026-01-11 9534c2c
Updated software versions in .bib file
