+++ article/paper.bib
@@ -97,12 +97,12 @@
   author    = {Białecki, Andrzej and
                Białecki, Piotr and
                Krupiński, Leszek},
-  title     = {{Kaszanas/DatasetPreparator: 2.0.0 
+  title     = {{Kaszanas/DatasetPreparator: 2.1.0 
                SC2DatasetPreparator Release}},
   month     = {jun},
   year      = {2022},
   publisher = {Zenodo},
-  version   = {2.0.0},
+  version   = {2.1.0},
   doi       = {10.5281/zenodo.5296664},
   url       = {https://doi.org/10.5281/zenodo.5296664}
 }
@@ -122,24 +122,24 @@
   author    = {Białecki, Andrzej and
                Krupiński, Leszek and
                Białecki, Piotr},
-  title     = {{Kaszanas/SC2InfoExtractorGo: 2.1.1 
+  title     = {{Kaszanas/SC2InfoExtractorGo: 2.1.3 
                SC2InfoExtractorGo Release}},
   month     = {jun},
   year      = {2022},
   publisher = {Zenodo},
-  version   = {2.1.1},
+  version   = {2.1.3},
   doi       = {10.5281/zenodo.5296788},
   url       = {https://doi.org/10.5281/zenodo.5296788}
 }
 @misc{Bialecki_2021_SC2AnonServerPy,
   author    = {Białecki, Andrzej and
                Białecki, Piotr},
-  title     = {{Kaszanas/SC2AnonServerPy: 1.0.0 SC2AnonServerPy 
+  title     = {{Kaszanas/SC2AnonServerPy: 1.1.0 SC2AnonServerPy 
                Release}},
   month     = {jul},
   year      = {2021},
   publisher = {Zenodo},
-  version   = {1.0.0},
+  version   = {1.1.0},
   doi       = {10.5281/zenodo.5138313},
   url       = {https://doi.org/10.5281/zenodo.5138313}
 }
