### 2026-01-11 Kaszanas/SC2Tools#2 @Kaszanas
@nkrusch this is done in #3
