### 2026-01-13 e678df2
Added information about SC2AnonServerPy, as per review
