### 2026-01-28 #8889 @DiegoAscanio
**Recommendation Update - Review Comment**

The authors have sufficiently addressed all aspects that previously rendered the manuscript unsuitable for publication.

The submodules are now properly documented, authors have included tests for each submodule, and the solution has been packaged in a Docker image, significantly enhancing usability and reproducibility. However, I noted that the links to the submodules' documentation in the main repository's `README.md` are currently broken.

Minor corrections are required before final acceptance: specifically, fixing the broken documentation links in the README and inserting a proper header for the References section in paper.md, as the citations are currently merging with the Acknowledgements.

Regarding the work's suitability for JOSS, it is worth noting that the authors have previously published related papers in high-impact journals, such as Scientific Data (Nature), and this submission represents a continuation of an established, peer-reviewed line of research.

Regarding the **'State of the Field'** checklist item: While this is typically a mandatory section, I agree with the authors' current presentation and do not believe an explicit comparison section is necessary in this specific case.

My reasoning is that, based on my investigation of the literature, there are no direct equivalent tools available. I conducted searches on Google Scholar using keywords such as 'Starcraft II dataset processing tool' and 'Starcraft II gameplay extraction tool'. Although there are over 22,000 papers related to StarCraft II, I could not find a comparable tool suite that handles dataset processing in this manner.

Therefore, requiring a **'State of the Field'** section to compare against non-existent or distantly related tools would add little value. The current context provided in the paper is sufficient."

**Conclusion:** After the authors fix the documentation links and the Reference section header in paper.md, I consider this paper suitable for publication at JOSS.

**AI use Acknowledgments:** I'm a non-native English speaker, so I admit the use of LLM tools - google Gemini 3 - to improve my **review comment** readability as the original one could have some hard to understand parts, but, I do share the original evaluation -  that I asked for google's Gemini LLM improvement - available at: https://gist.github.com/DiegoAscanio/57560d9912d029881eb2f81401f452fb
