### 2026-01-28 41530ea
Fixing README links
As requested in: https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3810168898
