### 2026-01-28 866774f
Fixing References heading
As requested by: https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3810168898
