### 2026-01-28 #8889 @Kaszanas
@nkrusch @DiegoAscanio Thank you very much for spotting these minor issues. Below you will find my response.

> However, I noted that the links to the submodules' documentation in the main repository's README.md are currently broken.

As of now, the README.md is fixed with links that point towards the proper files.

> inserting a proper header for the References section in paper.md, as the citations are currently merging with the Acknowledgements.

Proper header was inserted.

Thank you for your time and engagement in improving our work.
