### 2026-01-29 #8889 @nkrusch
Hi @DiegoAscanio,

In your [checklist](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3238386514), the item "performance" remains unchecked. Should the author make more changes?

The performance criteria is applicable when a paper makes explicit claims, such as empirical experiments with reported results. If no such claims are made, the criteria is automatically satisfactory.
