### 2026-02-03 #8889 @Kaszanas
> Hi [@DiegoAscanio](https://github.com/DiegoAscanio),
>
> In your [checklist](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3238386514), the item "performance" remains unchecked. Should the author make more changes?
>
> The performance criteria is applicable when a paper makes explicit claims, such as empirical experiments with reported results. If no such claims are made, the criteria is automatically satisfactory.

@nkrusch @DiegoAscanio @martinmodrak

If there are any more requests for article and software changes, please let me know. ;)
