### 2026-02-04 #8889 @nkrusch
@Kaszanas, on my part the situation looks like this (we need two complete reviews):

* @DiegoAscanio review is missing resolution on performance; the review is incomplete without it.
* Waiting for an update from @martinmodrak
