### 2026-02-08 #8889 @DiegoAscanio
Apologies for the oversight. I have now checked the item.

To clarify, the Performance criterion is not applicable in this context. The software is mainly a specialized tool for StarCraft II game evaluation and does not make specific performance claims (e.g., speed or efficiency improvements) over existing software. To the best of my knowledge, there are no direct comparable tools in the current literature to benchmark against.

So now it is checked (:

> Hi [@DiegoAscanio](https://github.com/DiegoAscanio),
>
> In your [checklist](https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3238386514), the item "performance" remains unchecked. Should the author make more changes?
>
> The performance criteria is applicable when a paper makes explicit claims, such as empirical experiments with reported results. If no such claims are made, the criteria is automatically satisfactory.
