### 2026-02-17 #8889 @martinmodrak
I finally got back to the review. I was able to install and run the tool per instructions.

I have some minor issues with the paper:

- Summary: Should probably end first paragraph/start second paragraph with saying this is a tool to process replays of Starcraft II (and possible some very short note what SC2 is)
- Summary: I don't think it is necessary to explain that tooling is useful
- Software description: "The full list of current submodules is as follows" ... the order here should probably match the order of the sections below
- There is some mismatch between Figure 1 and the text. I would suggest either having a list in text or in graphic format, but not necessarily both (unless there is something more to be learned from Figure 1 that is not in the text and which I am missing)

I also created two issues at the relevant repositories. Thouh the only one I think really _needs_ some care is https://github.com/Kaszanas/SC2_Datasets/issues/88
