### 2026-02-17 #8889 @martinmodrak
Also, one more formal thing is that (unless I am missing something) the [SC2AnonServerPy](https://github.com/Kaszanas/SC2AnonServerPy) and the [SC2Tools](https://github.com/Kaszanas/SC2Tools) lack a contributor guide.

Other than that, I think this is ready for publication.
