### 2026-02-17 1b92453
Adjusted summary as per review
https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3914842756
