### 2026-02-17 d356b85
enumeration follows sections
https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3914842756
