+++ paper.md
@@ -68,7 +68,7 @@ So far, our software was leveraged in preparation of major datasets: "SC2ReSet" 
 
 # Software Description
 
-Our software consists of multiple modules that the user can match to their specific needs. To easily extend our toolset, the main repository of "SC2Tools" contains multiple git submodules. Each submodule is a separate repository with the logic required to perform a specific tasks on the SC2Replay files. The full pipeline in a simplified pictorial form is showcased in \autoref{fig:DatasetPipeline}. The motivation for this structure is twofold. Firstly, it makes evolving the toolset easier, as modules can be easily replaced, or new ones added. Secondly, users have the option of using only a portion of the pipeline. The full list of current submodules is as follows: (1) "SC2InfoExtractorGo" [@Bialecki_2021_SC2InfoExtractorGo], (2) "DatasetPreparator" [@Bialecki_2022_SC2DatasetPreparator], (3) "SC2AnonServerPy" [@Bialecki_2021_SC2AnonServerPy]. (4) "SC2_Datasets" [@bialecki_2022_sc2datasets].
+Our software consists of multiple modules that the user can match to their specific needs. To easily extend our toolset, the main repository of "SC2Tools" contains multiple git submodules. Each submodule is a separate repository with the logic required to perform a specific tasks on the SC2Replay files. The full pipeline in a simplified pictorial form is showcased in \autoref{fig:DatasetPipeline}. The motivation for this structure is twofold. Firstly, it makes evolving the toolset easier, as modules can be easily replaced, or new ones added. Secondly, users have the option of using only a portion of the pipeline. The full list of current submodules is as follows: (1) "DatasetPreparator" [@Bialecki_2022_SC2DatasetPreparator], (2) "SC2InfoExtractorGo" [@Bialecki_2021_SC2InfoExtractorGo], (3) "SC2AnonServerPy" [@Bialecki_2021_SC2AnonServerPy]. (4) "SC2_Datasets" [@bialecki_2022_sc2datasets].
 
 ![Simplified full pipeline using SC2Tools to create two datasets, "SC2ReSet" [@Bialecki2022ReSetZenodo] and "SC2EGSet" Dataset [@Bialecki2023EGSetZenodo]. Initially introduced in [@Bialecki2023SC2EGSet]. \label{fig:DatasetPipeline}](./article/SC2Tools_Pipeline_Complete.pdf)
 
