### 2026-02-17 #8889 @Kaszanas
Hi @martinmodrak, please find the response and listed adjustments below. Thank you for the input.

> I finally got back to the review. I was able to install and run the tool per instructions.
>
> I have some minor issues with the paper:
>
> * Summary: Should probably end first paragraph/start second paragraph with saying this is a tool to process replays of Starcraft II (and possible some very short note what SC2 is)
> * Summary: I don't think it is necessary to explain that tooling is useful

The summary was adjusted, one sentence was removed. I have added a mention about what StarCraft 2 is. This cannot be more extensive due to the word limits.

> * Software description: "The full list of current submodules is as follows" ... the order here should probably match the order of the sections below

This was adjusted, the enumeration follows sections.

> * There is some mismatch between Figure 1 and the text. I would suggest either having a list in text or in graphic format, but not necessarily both (unless there is something more to be learned from Figure 1 that is not in the text and which I am missing)

If possible I would like to keep both. The pictorial representation itself is just an example of a full pipeline, the reader does not need to fully focus on all of the details in the article to have a grasp of the toolset. I am fine with removing it if it is absolutely necessary, but in my opinion it is a nice addition. Let me know which solution do you recommend.

> I also created two issues at the relevant repositories. Thouh the only one I think really _needs_ some care is [Kaszanas/SC2_Datasets#88](https://github.com/Kaszanas/SC2_Datasets/issues/88)

The documentation request was implemented in the SC2_Datasets repository, I hope that the provided information resolves the issue. Additionally, as we have discussed in the SC2Tools repository, the generated summaries may be subject to change in the future.

Once again thank you for engaging in the process and providing your feedback!

I will pin new versions of the software to the SC2Tools repository if these changes are deemed fully sufficient.
