### 2026-02-18 #8889 @martinmodrak
@Kaszanas - it appears `paper.md` was not updated on GitHub...
