### 2026-02-18 #8889 @Kaszanas
> @Kaszanas - it appears `paper.md` was not updated on GitHub...

Let me fix that, the commits were made on a branch.
