### 2026-02-18 #8889 @nkrusch
@martinmodrak - your comments are appreciated and encouraged. The subject reviewers insight is, in general, critical to spot for example overstatements and omissions, and naturally we want to me mindful of those. That said, you should continue until the state of the submission is satisfactory (I will review the paper separately after the reviewers).

I am also in favor of adding contribution guidelines at the top-level; particularly since that is the entry point to the implementation. If the submodules have varying guidelines, then the top-level guidelines can explain that and refer to the submodules.
