### 2026-02-18 dd3aec6
Dropped 3rd paragraph as per review
https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3920675862
