### 2026-02-18 a69e532
non-breaking spaces everywhere, adjusted summary
https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3920675862
