+++ paper.md
@@ -45,9 +45,9 @@ bibliography: ./article/paper.bib
 
 # Summmary
 
-Computer games, as fully controlled simulated environments, have been utilized in significant scientific studies demonstrating the application of Reinforcement Learning (RL). Gaming and esports are key areas influenced by the application of Artificial Intelligence (AI) and Machine Learning (ML) solutions at scale. Tooling simplifies scientific workloads and is essential for developing the gaming and esports research area.
+Computer games, as fully controlled simulated environments, have been utilized in significant scientific studies demonstrating the application of Reinforcement Learning (RL). Gaming and esports are key areas influenced by the application of Artificial Intelligence (AI) and Machine Learning (ML) solutions at scale. Tooling simplifies scientific workloads and is essential for developing the gaming and esports research area. Our tools primarily deliver software for working with StarCraft&nbsp;2, a well-known real-time strategy (RTS) game and one of the long-standing esports titles.
 
-In this work, we present "SC2Tools", a toolset containing multiple submodules responsible for working with, and producing larger datasets. StarCraft&nbsp;2 is a known real-time strategy game (RTS) and one of the long-standing esports titles. We provide a modular structure of the implemented tooling, leaving room for future extensions where needed. Additionally, some of the tools are not StarCraft&nbsp;2 exclusive and can be used with other types of data for dataset creation. We provide PyTorch and PyTorch Lightning application programming interface (API) for easy access to the data. Finally, our solution provides some foundational work toward normalizing experiment workflow in StarCraft&nbsp;2
+In this work, we present "SC2Tools", a toolset containing multiple submodules responsible for working with, and producing larger StarCraft&nbsp;2 datasets. We provide a modular structure of the implemented tooling, leaving room for future extensions where needed. Additionally, some of the tools are not StarCraft&nbsp;2 exclusive and can be used with other types of data for dataset creation. We provide PyTorch and PyTorch Lightning application programming interface (API) for easy access to the data. Finally, our solution provides some foundational work toward normalizing experiment workflow in StarCraft&nbsp;2
 
 # Introduction and Background
 
@@ -98,20 +98,20 @@ Users have the ability to extend our solution and apply it to their data via the
 
 ## Software Functionalities
 
-Main functionality of this software collection introduce a repeatable way of working with StarCraft 2 data for research and data analysis. Users need to verify if their specific use case is permitted by the Blizzard End User License Agreement (EULA). Our software package includes file-wrangling tools such as: flattening nested directory structure, data-parallel replay file parsing (extraction), data cleanup, exporting replay data to JSON, and finally data loading into PyTorch [@PyTorch2019] and PyTorch Lightning [@PyTorch_Lightning_2019]. We have developed a modular system of tools solving specific issues of data processing with expandability in mind.
+Main functionality of this software collection introduce a repeatable way of working with StarCraft&nbsp;2 data for research and data analysis. Users need to verify if their specific use case is permitted by the Blizzard End User License Agreement (EULA). Our software package includes file-wrangling tools such as: flattening nested directory structure, data-parallel replay file parsing (extraction), data cleanup, exporting replay data to JSON, and finally data loading into PyTorch [@PyTorch2019] and PyTorch Lightning [@PyTorch_Lightning_2019]. We have developed a modular system of tools solving specific issues of data processing with expandability in mind.
 
 
 # Statement of Need
 
[…]
