### 2026-02-18 0ea2fa0
Added CONTRIBUTING.md
https://github.com/openjournals/joss-reviews/issues/8889#issuecomment-3920675862
