### 2026-02-18 #8889 @Kaszanas
> **The Summary section** I fear the summary still reads a bit clunky - it is not misleading or anything, so I am not sure how much I should micromanage this and how much it is OK to just let it be. Presumably [@nkrusch](https://github.com/nkrusch) has an opinion and I am ready to accept the paper as is if they think I am overdoing this.
> Notably when you write
>
> > In this work, we present “SC2Tools”, a toolset  ... for working with, and producing larger datasets.
>
> It appears as if this is a general tool for any dataset, which it mostly isn't (and the reader will likely understand this once they go through the rest of the paragraph).
>
> It would make more sense to me to structure the summary roughly as:
>
> * Computer games are a thing for ML/AI research. SC2 is a popular RTS game.
> * We present a tool to work with SC2 replays, summary of the tool and the main design choices
>
> I think the current 3rd paragraph is mostly superfluous and has little information value (though it is not wrong), so I would consider just dropping it.

Thank you, these changes were applied, the sentence introducing StarCraft 2 now reads before the second paragraph. Third paragraph was partly dropped. please see the commit changes to paper.md.

> Also, you have "Starcraft~2" rendered in the PDF (presumably this was supposed to be a non-breaking space)

This was fixed, sorry, I am used to `~` for non-breaking spaces from LaTeX.

> **Other considerations**
>
> I am OK with keeping Figure 1 in, that definitely wasn't intended as a binding requirement.
>
> I am still missing some of the contributing guidelines.

I have added the `CONTRIBUTING.md` files to SC2Tools and SC2AnonServerPy. I originally omitted this as a mistake, because this information was not in the repository issues, sorry.

> Other than that the paper is IMHO ready for acceptance.

Thank you very much. I hope that the changes I have made in this latest response are well aligned with your requests.

@nkrusch In case of final changes, please let me know, I will update the submodule versions to the latest commits.
