### 2026-02-19 #8889 @martinmodrak
@Kaszanas thanks for the updates. I think the submission is in good enough state to be published in JOSS (@nkrusch)
