### 2026-02-19 #8889 @nkrusch
@Kaszanas please check these two bib entries; these DOIs do not exist.

> ❌ INVALID DOIs
>
> - 10.5555/3455716.3455894 is INVALID - 10.5555 is a known broken prefix, replace with https://dl.acm.org/doi/{doi} in the {url} field
> - 10.5555/3454287.3455008 is INVALID - 10.5555 is a known broken prefix, replace with https://dl.acm.org/doi/{doi} in the {url} field
