### 2026-02-19 #8889 @Kaszanas
> [@Kaszanas](https://github.com/Kaszanas) please check these two bib entries; these DOIs do not exist.
>
> > ❌ INVALID DOIs
> >
> > * 10.5555/3455716.3455894 is INVALID - 10.5555 is a known broken prefix, replace with [https://dl.acm.org/doi/{doi}](https://dl.acm.org/doi/%7Bdoi%7D) in the {url} field
> > * 10.5555/3454287.3455008 is INVALID - 10.5555 is a known broken prefix, replace with [https://dl.acm.org/doi/{doi}](https://dl.acm.org/doi/%7Bdoi%7D) in the {url} field

Thank you, I have fixed these. See the PR: https://github.com/Kaszanas/SC2Tools/pull/12
