### 2026-02-19 Kaszanas/SC2Tools#13 @nkrusch
Requested bib edits
Ref: [JOSS #8889](https://github.com/openjournals/joss-reviews/issues/8889)

* [ ] Name "Boxcars" does not match the repository name (awpy), and there is later a separate entry for apwy:
    >    Babcock, N. (2016). Boxcars. https://github.com/pnxenopoulos/awpy.
    >   Xenopoulos, P. (2020). Awpy. https://github.com/pnxenopoulos/awpy

* [ ] The Zenodo title should match the bib title:
    > Białecki, P. (2021). Kaszanas/SC2AnonServerPy: 1.1.0 SC2AnonServerPy

* [ ] The author list does not match between bib and Zenodo:
    > Białecki, A., Białecki, P., & Krupiński, L. (2022). Kaszanas/DatasetPreparator: 2.1.0
    > Białecki, A., Krupiński, L., & Białecki, P. (2022). Kaszanas/SC2InfoExtractorGo: 2.1.3

* [ ] Ideally, for repositories include a version if it exists:
    > Blizzard. (2017). s2client-proto. In GitHub. https://github.com/Blizzard/s2client-proto195
    > grpc. (2014). Grpc. https://github.com/grpc/grpc.h
    > Seeger, M. (2022). sc2-ai-coach. https://github.com/manuelseeger/sc2-ai-coach.
    > skadistats. (2013). Clarity. https://github.com/skadistats/clarity.

* [ ] Issue volume/number "0(0)" is wrong:
    > Brock, T. (2023). ... Philosophy, 0(0)

* [ ] Journal title capitalization here is inconsistent to other bib entries:
    > The science of preparing, enduring and winning, part c

* [ ] Add a "date accessed" because these are websites:
    > Chan, D. (2020). SC2 revealed. https://sc2revealed.com/.
    > Fonn, E. (2011). Aligulac. http://aligulac.com/.
    > Martin, A. (n.d.). sc2replaystats. https://sc2replaystats.com/.
    > Tool, S. (2013). Spawning tool.  https://lotv.spawningtool.com/

* [ ] Add year:
    > Martin, A. (n.d.). sc2replaystats. https://sc2replaystats.com/.

* [ ] The title says version 1.4, but DOI link points to version 0.7.6
    > Falcon, W., & The PyTorch Lightning team. (2019). PyTorch Lightning (Version 1.4).

* [ ] "This content has been removed by author" (is there a non-researchGate DOI?)
    > Bessi, S., ... https://doi.org/10.23913140/RG.2.2.10296.25605240

* [ ] Duplicated entry (L241/L245)
    > Kim, G., Joerg, D...

* [ ] The caps "ARXIV" stands out when other entries use "arXiv":
    > https://doi.org/10.48550/ARXIV.2011.00958309
