### 2026-02-19 Kaszanas/SC2Tools#14 @nkrusch
Paper content edits
Ref: [JOSS #8889](https://github.com/openjournals/joss-reviews/issues/8889)

You will want a parenthesized citation here, not a text cite:
> These attributes make for a great environment for testing various AI agents Vinyals et al. (2019).

The comma here seems out of place. However, changing it to a "." makes me wonder if it is best to just remove the first part? In any case, some rewording is needed here.
> The need for similar solutions is clear, output of this software or related artifacts were used...

Period (not comma)
> software to output JSON files, We claim full interoperability
