### 2026-02-19 #8889 @nkrusch
@Kaszanas - thanks I have left comments about the paper in two issues. Most comments are just about tidying up the bibliography; nothing major. Let me know when ready, and I will proceed.
