### 2026-02-19 baf502b
Fixing citations "," -> ";", rendering issue
https://github.com/Kaszanas/SC2Tools/issues/14#issue-3963900973
