### 2026-02-19 d6ef508
Changed comma to period. editing
https://github.com/Kaszanas/SC2Tools/issues/14#issue-3963900973
