### 2026-02-19 eeb8a08
comma -> period, editing
https://github.com/Kaszanas/SC2Tools/issues/14#issue-3963900973
