### 2026-02-19 Kaszanas/SC2Tools#14 @Kaszanas
> Ref: [JOSS #8889](https://github.com/openjournals/joss-reviews/issues/8889)
>
> You will want a parenthesized citation here, not a text cite:
>
> > These attributes make for a great environment for testing various AI agents Vinyals et al. (2019).

I have fixed this, one of the separators was `,` instead of `;`, an omission when changing the format from LaTeX (this resulted in missing citations):

The sentence now reads:
```
These attributes make for a great environment for testing various AI agents [@Ma2024LLMStarCraft2;@Vinyals2019;@Samvelyan2019SMAC;@Pearce2022CSGO].
```

> The comma here seems out of place. However, changing it to a "." makes me wonder if it is best to just remove the first part? In any case, some rewording is needed here.
>
> > The need for similar solutions is clear, output of this software or related artifacts were used...

Switched from comma to a period. I am open to removing the first part if needed, although it offers a strong claim that later is backed up by the following paragraphs.

> Period (not comma)
>
> > software to output JSON files, We claim full interoperability

Adjusted.

PR: https://github.com/Kaszanas/SC2Tools/pull/15

Thank you for spotting these issues.
