### 2026-02-19 30a93e9
adjusted boxcars url field.
https://github.com/Kaszanas/SC2Tools/issues/13
