+++ article/paper.bib
@@ -574,7 +574,7 @@
 @misc{URLBoxcars2016,
   author       = {Nick Babcock},
   title        = {boxcars},
-  howpublished = {\url{https://github.com/pnxenopoulos/awpy}},
+  howpublished = {\url{https://github.com/nickbabcock/boxcars}},
   year         = {2016},
   note         = {Acessed: 2024.09.20}
 }
