### 2026-02-19 3bd232b
repos with versions, verified author lists
https://github.com/Kaszanas/SC2Tools/issues/13
