+++ article/paper.bib
@@ -120,8 +120,9 @@
 }
 @misc{Bialecki_2021_SC2InfoExtractorGo,
   author    = {Białecki, Andrzej and
-               Krupiński, Leszek and
-               Białecki, Piotr},
+               Białecki, Piotr and
+               Krupiński, Leszek
+               },
   title     = {{Kaszanas/SC2InfoExtractorGo: 2.1.3 
                SC2InfoExtractorGo Release}},
   month     = {jun},
@@ -465,7 +466,7 @@
 }
 @misc{URLS2Prot2016,
   author       = {András Belicza},
-  title        = {s2prot},
+  title        = {s2prot v1.5.2},
   howpublished = {\url{https://github.com/icza/s2prot}},
   year         = {2016},
   note         = {Acessed: 2024.10.12}
@@ -479,14 +480,14 @@
 }
 @misc{URLSC2AICoach2022LLM,
   author       = {Manuel Seeger},
-  title        = {sc2-ai-coach},
+  title        = {sc2-ai-coach v0.5.5},
   howpublished = {\url{https://github.com/manuelseeger/sc2-ai-coach}},
   year         = {2022},
   note         = {Acessed: 2024.10.12}
 }
 @misc{gRPC2014,
   author       = {grpc},
-  title        = {grpc},
+  title        = {grpc v1.78.0},
   howpublished = {\url{https://github.com/grpc/grpc}},
   year         = {2014},
   note         = {Acessed: 2025.03.14}
@@ -566,14 +567,14 @@
 
 @misc{AWPYXeno2020,
   author       = {Peter Xenopoulos},
-  title        = {awpy},
+  title        = {{awpy 2.0.2}},
   howpublished = {\url{https://github.com/pnxenopoulos/awpy}},
   year         = {2020},
   note         = {Acessed: 2024.09.20}
 }
 @misc{URLBoxcars2016,
   author       = {Nick Babcock},
-  title        = {boxcars},
+  title        = {{boxcars v0.10.10}},
   howpublished = {\url{https://github.com/nickbabcock/boxcars}},
   year         = {2016},
   note         = {Acessed: 2024.09.20}
@@ -581,7 +582,7 @@
 @misc{URLBlizzardS2ClientProto,
   url     = {https://github.com/Blizzard/s2client-proto},
   author  = {Blizzard},
-  title   = {{s2client-proto}},
+  title   = {{s2client-proto 1.2.0}},
   journal = {GitHub},
   year    = {2017},
   note    = {Acessed: 2024.09.20}
@@ -638,14 +639,14 @@
 }
 @misc{OpenDotaGitHub,
   author       = {odota},
-  title        = {core},
+  title        = {{core}},
   howpublished = {\url{https://github.com/odota/core}},
   year         = {2014},
   note         = {Acessed: 2025.01.01}
 }
 @misc{ClarityGitHub,
   author       = {skadistats},
-  title        = {clarity},
+  title        = {{clarity 2.1}},
   howpublished = {\url{https://github.com/skadistats/clarity}},
   year         = {2013},
   note         = {Acessed: 2025.01.01}
@@ -668,14 +669,14 @@
 }
 @misc{URLSc2replaystats,
   author       = {Anthony Martin},
-  title        = {sc2replaystats},
+  title        = {{sc2replaystats}},
   howpublished = {\url{https://sc2replaystats.com/}},
   note         = {Acessed: 2025.01.01}
 }
 @misc{URLSpawningTool,
   author       = {Spawning Tool},
   howpublished = {\url{https://lotv.spawningtool.com/}},
-  title        = {Spawning Tool},
+  title        = {{Spawning Tool}},
   year         = {2013},
   month        = {05},
[…]
