### 2026-02-19 e553443
Verified the Tom Brock entry, fixed
https://github.com/Kaszanas/SC2Tools/issues/13
