+++ article/paper.bib
@@ -697,16 +697,14 @@
   month        = {04},
   note         = {Accessed: 2025.03.17}
 }
-
-
 @article{Brock2023Interdisciplinary,
   author    = {Tom Brock},
-  title     = {{Ontology and interdisciplinary research in esports}},
+  title     = {Ontology and interdisciplinary research in esports},
   journal   = {Sport, Ethics and Philosophy},
-  volume    = {0},
-  number    = {0},
-  pages     = {1--17},
-  year      = {2023},
+  volume    = {19},
+  number    = {1},
+  pages     = {48--64},
+  year      = {2025},
   publisher = {Routledge},
   doi       = {10.1080/17511321.2023.2260567},
   url       = {https://doi.org/10.1080/17511321.2023.2260567},
