### 2026-02-19 1153e94
Fixing booktitle capitalization
https://github.com/Kaszanas/SC2Tools/issues/13
