+++ article/paper.bib
@@ -214,7 +214,7 @@
   volume    = {240},
   pages     = {161-174},
   year      = {2018},
-  booktitle = {Sport and the Brain: The Science of Preparing, Enduring and Winning, Part C},
+  booktitle = {{Sport and the Brain: The Science of Preparing, Enduring and Winning, Part C}},
   issn      = {0079-6123},
   doi       = {10.1016/bs.pbr.2018.09.006},
   url       = {https://doi.org/10.1016/bs.pbr.2018.09.006},
