### 2026-02-19 32d62ae
sc2replaystats year, lightning newest version, removed -> added URL
https://github.com/Kaszanas/SC2Tools/issues/13
