+++ article/paper.bib
@@ -447,12 +447,12 @@
 }
 @misc{PyTorch_Lightning_2019,
   author  = {Falcon, William and {The PyTorch Lightning team}},
-  doi     = {10.5281/zenodo.3828935},
+  doi     = {10.5281/zenodo.3530844},
   license = {Apache-2.0},
   month   = {3},
   title   = {{PyTorch Lightning}},
   url     = {https://github.com/PyTorchLightning/pytorch-lightning},
-  version = {1.4},
+  version = {2.6.1},
   year    = {2019}
 }
 
@@ -670,6 +670,7 @@
 @misc{URLSc2replaystats,
   author       = {Anthony Martin},
   title        = {{sc2replaystats}},
+  year         = {2012},
   howpublished = {\url{https://sc2replaystats.com/}},
   note         = {Acessed: 2025.01.01}
 }
@@ -932,7 +933,7 @@
   month  = {03},
   pages  = {},
   title  = {European report: Mental health outcomes in esports players},
-  doi    = {10.13140/RG.2.2.10296.25605}
+  url    = {https://www.researchgate.net/publication/389859780_European_report_Mental_health_outcomes_in_esports_players}
 }
 @article{Dupuy2025,
   author  = {Dupuy, Antoine and Campbell, Mark and Toth, Adam},
