### 2026-02-19 47d54a0
Removed duplicate sc2reader entry
https://github.com/Kaszanas/SC2Tools/issues/13
