+++ article/paper.bib
@@ -165,47 +165,6 @@
   doi     = {10.1038/s41597-023-02510-7},
   url     = {https://doi.org/10.1038/s41597-023-02510-7}
 }
-
-@software{ZenodoSC2Reader,
-  author    = {Graylin Kim and
-               David Joerg and
-               Kevin Leung and
-               Alexander Hanhikoski and
-               Christian Clauss and
-               Rasmus Précenth and
-               Dominik Neise and
-               Hugo Wainwright and
-               Cameron Zemek and
-               Andrene and
-               srounet and
-               Issac Kelly and
-               Jonathan Chung and
-               Bean Deng and
-               Talv and
-               Chazzz and
-               Jonathan Gravel and
-               rejuxst and
-               Anders Nickelsen and
-               Gusgus01 and
-               Dan Fulton and
-               DasFranck and
-               Rohit Sanbhadti and
-               Spencer Krum and
-               Tim Gates and
-               Bas Peschier and
-               Chris Lundquist and
-               David Kuhta and
-               eqy and
-               Kanwei Li},
-  title     = {{ggtracker/sc2reader: v1.8.0 Various fixes and 
-               improvements}},
-  month     = {05},
-  year      = {2022},
-  publisher = {Zenodo},
-  version   = {v1.8.0},
-  doi       = {10.5281/zenodo.6519543},
-  url       = {https://doi.org/10.5281/zenodo.6519543}
-}
 @incollection{Campbell2018,
   title     = {{Chapter 10 - eSports: A new window on neurocognitive expertise?}},
   editor    = {Samuele Marcora and Mustafa Sarkar},
