### 2026-02-19 70050d3
ARXIV -> arXiv
https://github.com/Kaszanas/SC2Tools/issues/13
