+++ article/paper.bib
@@ -348,7 +348,7 @@
   url     = {https://doi.org/10.1109/JIOT.2021.3074740}
 }
 @misc{SmerdovLoLDataset2021,
-  doi       = {10.48550/ARXIV.2011.00958},
+  doi       = {10.48550/arXiv.2011.00958},
   url       = {https://arxiv.org/abs/2011.00958},
   author    = {Smerdov, Anton and Zhou, Bo and Lukowicz, Paul and Somov, Andrey},
   keywords  = {Human-Computer Interaction (cs.HC), Artificial Intelligence (cs.AI), Machine Learning (cs.LG), FOS: Computer and information sciences, FOS: Computer and information sciences},
