### 2026-02-20 #8889 @Kaszanas
> [@Kaszanas](https://github.com/Kaszanas) Go ahead and merge the changes.
>
> After merging the PRs, at this time I also request that you complete the following steps.
>
> 1. Make a (new) tagged release of the software.
> 2. Archive the reviewed software in Zenodo, Figshare, or similar archival service.
> 3. Check the archival deposit has the correct metadata. This includes:
>
>    * title (should match the paper)
>    * author list (should match the paper)
>    * authors' ORCID (should match the paper)
>    * license (should match the repository)
>    * release version tag should match the archive version name
>      For example, git tag v1.0 would have archive name v1.0 (displays as "Version v1.0" on Zenodo).
> 4. List here, as a comment:
>
>    * the version tag
>    * the archive DOI
>
> I can then move forward with recommending accept for the submission.
>
> Since you have multi licenses, I recommend trying to reflect that in the Zenodo deposit. We will need the repository and Zenodo to have the same information.

I have merged the changes, note that the "Accessed on: <date>" notes were in the .bib file, but you stated these did not render to the .pdf. Let me know if that issue persists. I am not entirely sure how to solve this as this is mostly on the compilation side.

I will perform the rest of the instructions and let you know as soon as I do.
