### 2026-02-21 #8889 @Kaszanas
@nkrusch

Zenodo archive link: https://zenodo.org/records/18726432
doi: [10.5281/zenodo.18726431](https://doi.org/10.5281/zenodo.18726431)
version tag: 1.1.0

Everything should be in order, if it is not I can still adjust the metadata on the zenodo realease.
