### 2026-02-23 #8889 @Kaszanas
> [@Kaszanas](https://github.com/Kaszanas) Thanks!
>
> I see you added CC 4.0 to the top level. However, that is not an OSI license. I recommend instead you apply GPL-3.0 license at the top-level repository, because as that is used by the two submodules, and when mixing MIT and GPL-3.0, the subsequent work should be under GPL.
>
> On Zenodo, you just make a meta data change to change the license + make a new commit on GitHub.

Done.
