### 2026-02-23 #8889 @nkrusch
In the paper.md,  at L108, there is a reference to `ZenodoSC2Reader`.

I believe this should be `GitHubSC2Reader` (referring to Kim et al. "ggtracker/sc2reader").

It is this reference that is breaking.

Please make an update and I will retry.
