### 2026-02-23 64db4d9
ZenodoSC2Reader -> GitHubSC2Reader
