### 2026-02-23 #8889 @nkrusch
At this point wait for information from the track editor.
