### 2026-02-23 d36eb01
Add CITATION.cff for project citation information
Added citation information for the SC2Tools project, including authors, DOI, and preferred citation details.
