### 2026-02-23 638451f
Add JOSS status badge to README
Added JOSS status badge to README.
